+++ front-end/src/components/layoutComponents/Body/index.jsx
@@ -3,16 +3,16 @@ import { Routes, Route } from "react-router-dom";
 import "./style.scss";
 import PostPage from "../../pageComponents/PostsPage";
 import UsersPage from "../../pageComponents/UsersPage";
-import Current from "../../partsComponents/Current";
+// import Current from "../../partsComponents/Current";
 import NotFind from "../../pageComponents/NotFindPage";
 import Dashboard from "../../pageComponents/DashboardPage";
 
 export default function Body() {
   return (
     <div className="Body">
-      <Current />
+      {/* <Current /> */}
       <Routes>
-        <Route path="/" element={<PostPage />} />
+        <Route path="/" element={<Dashboard />} />
         <Route path="/Dashboard" element={<Dashboard />} />
         <Route path="/Posts" element={<PostPage />} />
         <Route path="/Users" element={<UsersPage />} />
