+++ front-end/src/index.js
@@ -1,14 +1,40 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import './index.css';
-import App from './App';
-import reportWebVitals from './reportWebVitals';
+import React from "react";
+import ReactDOM from "react-dom";
+import "./index.css";
+import reportWebVitals from "./reportWebVitals";
+import { BrowserRouter, Routes, Route } from "react-router-dom";
+import { store } from "./store";
+import { Provider } from "react-redux";
+import App from "./App";
+import Login from "./components/layoutComponents/Login";
+import { createTheme, ThemeProvider } from "@mui/material/styles";
+import { green, pink } from "@mui/material/colors";
+
+const myTheme = createTheme({
+  palette: {
+    primary: {
+      main: green[600],
+    },
+    secondary: {
+      main: pink[100],
+    },
+  },
+});
 
 ReactDOM.render(
   <React.StrictMode>
-    <App />
+    <Provider store={store}>
+      <ThemeProvider theme={myTheme}>
+        <BrowserRouter>
+          <Routes>
+            <Route path="/" element={<App />} />
+            <Route path="/login" element={<Login />} />
+          </Routes>
+        </BrowserRouter>
+      </ThemeProvider>
+    </Provider>
   </React.StrictMode>,
-  document.getElementById('root')
+  document.getElementById("root")
 );
 
 // If you want to start measuring performance in your app, pass a function
